+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Error and Warning messages
   - Improve error message for conflicting acquisition protocols (!982)
   - Fix overlapping operations warning due to floating point rounding errors (!989)
+  - Remove warning when latency corrections are not on a 4ns timegrid (!994)
 - Qblox backend
   - Change update param insertion logic. (!980)
   - Add compilation of `MarkerPulse` to `set_digital` via the `DigitalPulseStrategy` for QTM modules. (!975)